Move content_state global state to rarch_state
